[AWIT-130] code refactor - classes and headers moved to constants.ts, remove outsideClass from content json
