+++ frontend/src/utils/constants.ts
@@ -1,3 +1,5 @@
+import {ForumThread} from "../api";
+
 export const PageRoutes = {
     LOGIN: '/home',
     DASHBOARD: '/dashboard',
@@ -7,3 +9,31 @@ export const PageRoutes = {
     SETTINGS: '/settings',
     FORUM_THREAD: '/forum/thread'
 };
+
+const headers = ['Tytuł tematu', 'Liczba odpowiedzi', 'Założyciel', 'Data założenia','Akcje'];
+const classes = ['title', 'num', 'username', 'date', 'action'];
+
+const content = (thread: ForumThread) => [
+    {
+        outsideClass: classes[0],
+        content: thread.title
+    },
+    {
+        outsideClass: classes[1],
+        content: thread.forumPosts.length
+    },
+    {
+        outsideClass: classes[2],
+        content: thread.creator.username
+    },
+    {
+        outsideClass: classes[3],
+        content: thread.dateCreated.getFullYear()+"-"+(thread.dateCreated.getMonth()+1)+"-"+thread.dateCreated.getDate()
+    },
+    {
+        outsideClass: classes[4],
+        content: null
+    }
+]
+
+export {headers, classes, content}
